+++ packages/convergence/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- full support for returning promises in `.do()` blocks
+- timeout error thrown when convergence exceeds timeout
+
 ## [0.4.0] - 2018-03-02
 
 ### Changed